fix: fix only show neighbors on player response #78
